requirements.txt
astroid==2.15.5
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.0.12
dill==0.3.6
flake8==6.0.0
googlemaps==4.6.0
idna==3.3
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
platformdirs==3.5.1
pyaml==21.10.1
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.17.4
PyYAML==6.0
requests==2.27.1
tomlkit==0.11.8
typing_extensions==4.6.3
urllib3==1.26.9
wrapt==1.15.0
